Blitting now works
Texture2D has been updated to work with new texture map. Now text map
stores Texture2Ds and gives back pointers to whole class. Point has
also been updated to hold floats and not ints. All 6 drawing functions
now support blitting of images.

+++ include/GameFramework/Point.h
@@ -12,11 +12,11 @@ namespace GameFramework
 	class Point
 	{
 		public:
-			int X;
-			int Y;
+			float X;
+			float Y;
 
 			Point();
-			Point(int X, int Y);
+			Point(float X, float Y);
 	};
 }
